Retro Rabbit Tender
I found the idea behind the project to be highly intriguing. The underlying functioning of networks that the project wishes to expose to the user is something that we have handled in our lectures, so we have context for it. I think the project will be very interesting to work on and I believe that Amazon will provide invaluable experience. I understand that Amazon is expanding to South-Africa very rapidly, so I wish to partake in this ever growing company's affairs.